compile.c: suppress a warning
* compile.c (compile_data_alloc): cast to suppress a warning by
  apple gcc-4.2.  it doesn't cause an error but a warning in spite
  of -Werror=shorten-64-to-32 option, for some reason.

git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@46996 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ compile.c
@@ -612,7 +612,7 @@ compile_data_alloc(rb_iseq_t *iseq, size_t size)
     }
 
     ptr = (void *)&storage->buff[storage->pos];
-    storage->pos += size;
+    storage->pos += (int)size;
     return ptr;
 }
